Completed the addCourse and deleteCourse methods.
addCourse takes a user-inputted string from the searchfield and places it into a list containing all added courses.

deleteCourse takes the most recent addition into the course list and removes it (essentially popping).

In terms of collaboration with the UI, the added courses will be displayed in a label, and deleted courses will be removed from said label.

Next commit will focus on validating the input (ensuring they are legible). Thus, this commit assumes that all inputs are desirable inputs.